+++ src/components/content/condition-history/conditions-history.tsx
@@ -4,21 +4,18 @@ import {
   CollapsibleDataListStack,
   CollapsibleDataListStackEntries,
 } from "../../core/collapsible-data-list-stack";
-import { CCDAModal } from "../../core/modal-ccda";
 import { useCTW } from "../../core/providers/ctw-provider";
 import { DocumentButton } from "../CCDA/document-button";
+import { useCCDAModal } from "../CCDA/modal-ccda";
 import { ConditionHeader } from "../condition-header";
 import { applyConditionHistoryFilters } from "./condition-history-filters";
 import { conditionData, setupData } from "./condition-history-schema";
 import { Loading } from "@/components/core/loading";
+import { getBinaryId } from "@/fhir/binaries";
 import { getIncludedResources } from "@/fhir/bundle";
-import {
-  getBinaryDocument,
-  getBinaryId,
-  getProvenanceForConditions,
-  useConditionHistory,
-} from "@/fhir/conditions";
+import { useConditionHistory } from "@/fhir/conditions";
 import { ConditionModel } from "@/fhir/models/condition";
+import { searchProvenances } from "@/fhir/provenance";
 
 const CONDITION_HISTORY_LIMIT = 10;
 
@@ -39,30 +36,10 @@ export function ConditionHistory({
   const [conditionsWithoutDate, setConditionsWithoutDate] =
     useState<CollapsibleDataListStackEntries>([]);
   const [isHistoryLoading, setIsHistoryLoading] = useState(true);
-  const [binaryDocument, setBinaryDocument] = useState<fhir4.Binary>();
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [ccdaViewerTitle, setCCDAViewerTitle] = useState<string>();
 
   // Fetching
   const { getRequestContext } = useCTW();
   const historyResponse = useConditionHistory(condition);
-
-  // Handlers
-  const openCCDAModal = async (
-    binaryId: string,
-    conditionDisplayName?: string | undefined
-  ) => {
-    const requestContext = await getRequestContext();
-    const binaryDocumentResponse = await getBinaryDocument(
-      requestContext,
-      binaryId
-    );
-    setBinaryDocument(binaryDocumentResponse);
-    setIsModalOpen(true);
-    if (conditionDisplayName) {
-      setCCDAViewerTitle(conditionDisplayName);
-    }
-  };
 
   useEffect(() => {
     async function load() {
@@ -77,15 +54,14 @@ export function ConditionHistory({
         );
 
         const requestContext = await getRequestContext();
-        const provenanceBundles = await getProvenanceForConditions(
-          requestContext,
-          [condition, ...conditionsDataDeduped]
-        );
+        const provenances = await searchProvenances(requestContext, [
+          condition,
+          ...conditionsDataDeduped,
+        ]);
 
-        let binaryId;
         const conditionsToDisplay = conditionsDataDeduped.map(
           (dedupdedCondition) => {
-            binaryId = getBinaryId(provenanceBundles, dedupdedCondition.id);
+            const binaryId = getBinaryId(provenances, dedupdedCondition.id);
 
             return {
               ...setupData(dedupdedCondition),
@@ -112,32 +88,22 @@ export function ConditionHistory({
   }, [condition, getRequestContext, historyResponse.data, onEdit]);
 
   return (
-    <>
-      <CCDAModal
-        isOpen={isModalOpen}
-        fileName={ccdaViewerTitle}
-        rawBinary={binaryDocument}
-        onClose={() => setIsModalOpen(false)}
+    <div className="ctw-space-y-6">
+      <ConditionHeader condition={condition} />
+      <Details
+        data={conditionData(condition)}
+        readOnly={!onEdit}
+        onEdit={() => {
+          onClose();
+          onEdit?.();
+        }}
       />
-
-      <div className="ctw-space-y-6">
-        <ConditionHeader condition={condition} />
-        <Details
-          data={conditionData(condition)}
-          readOnly={!onEdit}
-          onEdit={() => {
-            onClose();
-            onEdit?.();
-          }}
-        />
-        <HistoryRecords
-          conditionsWithDate={conditionsWithDate}
-          conditionsWithoutDate={conditionsWithoutDate}
-          historyIsLoading={isHistoryLoading}
-          openCCDAModal={openCCDAModal}
-        />
-      </div>
-    </>
+      <HistoryRecords
+        conditionsWithDate={conditionsWithDate}
+        conditionsWithoutDate={conditionsWithoutDate}
+        historyIsLoading={isHistoryLoading}
+      />
+    </div>
   );
 }
 
@@ -145,18 +111,15 @@ type HistoryRecordsProps = {
   conditionsWithDate: CollapsibleDataListStackEntries;
   conditionsWithoutDate: CollapsibleDataListStackEntries;
   historyIsLoading: boolean;
-  openCCDAModal: (
-    binaryId: string,
-    conditionTitle: string | undefined
-  ) => Promise<void>;
 };
 
 const HistoryRecords = ({
   conditionsWithDate,
   conditionsWithoutDate,
   historyIsLoading,
-  openCCDAModal,
 }: HistoryRecordsProps) => {
+  const openCCDAModal = useCCDAModal();
+
   if (
     conditionsWithDate.length === 0 &&
     conditionsWithoutDate.length === 0 &&
@@ -168,45 +131,41 @@ const HistoryRecords = ({
     return <Loading message="Loading condition history..." />;
   }
 
+  function maybeAddDocumentButtons(entries: CollapsibleDataListStackEntries) {
+    return entries.map((entry) => {
+      if (!entry.binaryId) {
+        return entry;
+      }
+
+      return {
+        ...entry,
+        documentButton: (
+          <DocumentButton
+            onClick={() =>
+              openCCDAModal(
+                entry.binaryId as string,
+                entry.title ?? "Condition"
+              )
+            }
+            text="Source Document"
+          />
+        ),
+      };
+    });
+  }
+
   return (
     <>
       <CollapsibleDataListStack
-        entries={conditionsWithDate.map((entry) => ({
-          ...entry,
-          documentButton: (
-            <>
-              {entry.binaryId && (
-                <DocumentButton
-                  onClick={() =>
-                    openCCDAModal(entry.binaryId as string, entry.title)
-                  }
-                  text="Source Document"
-                />
-              )}
-            </>
-          ),
-        }))}
+        entries={maybeAddDocumentButtons(conditionsWithDate)}
         limit={CONDITION_HISTORY_LIMIT}
       />
+
       {conditionsWithoutDate.length !== 0 && (
         <div className="ctw-space-y-2">
           <div className="ctw-font-medium">Records with no date:</div>
           <CollapsibleDataListStack
-            entries={conditionsWithoutDate.map((entry) => ({
-              ...entry,
-              documentButton: (
-                <>
-                  {entry.binaryId && (
-                    <DocumentButton
-                      onClick={() =>
-                        openCCDAModal(entry.binaryId as string, entry.title)
-                      }
-                      text="Source Document"
-                    />
-                  )}
-                </>
-              ),
-            }))}
+            entries={maybeAddDocumentButtons(conditionsWithoutDate)}
             limit={CONDITION_HISTORY_LIMIT}
           />
         </div>